feat(formatter): dprint as astro, css, html, markdown, etc, formatter
dprint (https://dprint.dev/) supports astro and the like formatting but
not linting. We should run this before running biome; for in our repo
biome is configured in a non-standard way. This is because we imported
previous eslint + prettier config; and althought biome aims to be a
drop-in replacement for both -- and mostly, succeeds --, we just got a
tangle of legacy code mess.

Ideally, we should run both dprint and biome as vanilla as
possible. Moreover, we should aim to integrate biome to dprint.